music_mad.c: better skipping of tags:
Instead of doing it 'as-we-go' every time, skip the tags at file-open
time only once.

The former implementation had the chance of not having the full tag in
the frame, that includes a possibility of even not having the complete
tag magic in the frame leading to not identifying a tag.

The new implementation skips all the tags at file-start and file-end,
and does that only once. That requires some local file data-start and
file-end bookkeeping for SDL_RWops which are added.

Also added are detection and skipping of extended ID3v1 and APEv1 tags.

--HG--
branch : SDL-1.2